SA-control-3-main/requirements.txt
fastapi[all]
uvicorn[standard]
python-multipart
pyjwt
bcrypt
python-dotenv
slowapi